requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LLM and Agent
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
langchain-openai
langchain-ollama
langchain-core

# Observability
opik>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase

# Streaming
sse-starlette>=1.6.0

# Email
resend>=0.8.0

# Google Calendar
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
